Created pygameTest
Set up some code in the first few lines so that I can
test small pieces of code and then have it exit().
Learning note: I'm using control + z to background
the program so that I can run it from the command line.
Then I type fg or fg 1 to move it to the foreground again.